chore: no more v2 for document related services
