+++ bc_obps/service/document_service.py
@@ -1,17 +1,22 @@
-from typing import Optional, Tuple
+from typing import Tuple
 from uuid import UUID
+from service.data_access_service.document_service_v2 import DocumentDataAccessServiceV2
 from service.data_access_service.document_service import DocumentDataAccessService
-from registration.models import Document, DocumentType, Operation
-from service.data_access_service.operation_service import OperationDataAccessService
+from registration.models import Document, Operation
 from registration.utils import files_have_same_hash
 from django.core.files.base import ContentFile
+from service.data_access_service.operation_service import OperationDataAccessService
 
 
 class DocumentService:
     @classmethod
-    def get_existing_statutory_declaration_by_operation_id(cls, operation_id: UUID) -> Optional[Document]:
-        operation: Operation = OperationDataAccessService.get_by_id(operation_id)
-        return operation.documents.filter(type=DocumentType.objects.get(name="signed_statutory_declaration")).first()
+    def get_operation_document_by_type_if_authorized(
+        cls, user_guid: UUID, operation_id: UUID, document_type: str
+    ) -> Document | None:
+        from service.operation_service_v2 import OperationServiceV2
+
+        OperationServiceV2.get_if_authorized_v2(user_guid, operation_id, ['id', 'operator_id'])
+        return DocumentDataAccessService.get_operation_document_by_type(operation_id, document_type)
 
     @classmethod
     def create_or_replace_operation_document(
@@ -21,9 +26,9 @@ class DocumentService:
         This function receives a document and operation id.
         Operations only have one of each type of document, so this function uses the type to check if an existing document needs to be replaced, or if no document exists and one must be created.
         This function does NOT set any m2m relationships.
-        :returns: Tuple[Document, bool] where the bool is True if a new document was created, False if an existing document was updated
+        :returns:c Tuple[Document, bool] where the bool is True if a new document was created, False if an existing document was updated
         """
-        existing_document = DocumentDataAccessService.get_operation_document_by_type(operation_id, document_type)
+        existing_document = cls.get_operation_document_by_type_if_authorized(user_guid, operation_id, document_type)
         # if there is an existing  document, check if the new one is different
         if existing_document:
             # We need to check if the file has changed, if it has, we need to delete the old one and create a new one
@@ -32,7 +37,7 @@ class DocumentService:
             else:
                 return existing_document, False
         # if there is no existing document, create a new one
-        document = DocumentDataAccessService.create_document(user_guid, file_data, document_type)
+        document = DocumentDataAccessServiceV2.create_document(user_guid, file_data, document_type, operation_id)
         return document, True
 
     @classmethod
