+++ bc_obps/service/tests/test_document_service.py
@@ -1,4 +1,4 @@
-from service.data_access_service.document_service import DocumentDataAccessService
+from service.data_access_service.document_service_v2 import DocumentDataAccessServiceV2
 from registration.utils import data_url_to_file
 from registration.models.document import Document
 from registration.models.operation import Operation
@@ -11,7 +11,25 @@ from model_bakery import baker
 pytestmark = pytest.mark.django_db
 
 
-class TestDocumentService:
+class TestDocumentServiceV2:
+    @staticmethod
+    def test_get_operation_document_by_type_if_authorized():
+        approved_user_operator = baker.make_recipe('registration.tests.utils.approved_user_operator')
+        operation = baker.make_recipe('registration.tests.utils.operation', operator=approved_user_operator.operator)
+        document = baker.make_recipe('registration.tests.utils.document', operation=operation)
+        retrieved_document = DocumentService.get_operation_document_by_type_if_authorized(
+            approved_user_operator.user.user_guid, operation.id, 'boundary_map'
+        )
+        assert document.id == retrieved_document.id
+
+    @staticmethod
+    def test_cannot_get_operation_document_by_type_if_unauthorized():
+        user = baker.make_recipe('registration.tests.utils.industry_operator_user')
+        operation = baker.make_recipe('registration.tests.utils.operation')
+        baker.make_recipe('registration.tests.utils.document')
+        with pytest.raises(Exception, match='Unauthorized.'):
+            DocumentService.get_operation_document_by_type_if_authorized(user.user_guid, operation.id, 'boundary_map')
+
     @staticmethod
     def test_create_operation_document():
         # the value received by the service is a File (transformed into this in the django ninja schema)
@@ -30,10 +48,9 @@ class TestDocumentService:
     def test_do_not_update_duplicate_operation_document():
         approved_user_operator = baker.make_recipe('registration.tests.utils.approved_user_operator')
         operation = baker.make_recipe('registration.tests.utils.operation', operator=approved_user_operator.operator)
-        existing_document = DocumentDataAccessService.create_document(
-            approved_user_operator.user_id, data_url_to_file(MOCK_DATA_URL), 'boundary_map'
+        DocumentDataAccessServiceV2.create_document(
+            approved_user_operator.user_id, data_url_to_file(MOCK_DATA_URL), 'boundary_map', operation.id
         )
-        operation.documents.set([existing_document])
         created_at = operation.documents.first().created_at
 
         updated_file = data_url_to_file(MOCK_DATA_URL)
@@ -52,10 +69,9 @@ class TestDocumentService:
     def test_update_operation_document():
         approved_user_operator = baker.make_recipe('registration.tests.utils.approved_user_operator')
         operation = baker.make_recipe('registration.tests.utils.operation', operator=approved_user_operator.operator)
-        existing_document = DocumentDataAccessService.create_document(
-            approved_user_operator.user_id, data_url_to_file(MOCK_DATA_URL), 'boundary_map'
+        DocumentDataAccessServiceV2.create_document(
+            approved_user_operator.user_id, data_url_to_file(MOCK_DATA_URL), 'boundary_map', operation.id
         )
-        operation.documents.set([existing_document])
 
         updated_file = data_url_to_file(MOCK_DATA_URL_2)
         document, created = DocumentService.create_or_replace_operation_document(
@@ -74,13 +90,14 @@ class TestDocumentService:
         operation = baker.make_recipe(
             'registration.tests.utils.operation', operator=approved_user_operator.operator, status=registration_status
         )
-        b_map = DocumentDataAccessService.create_document(
-            approved_user_operator.user_id, data_url_to_file(MOCK_DATA_URL), 'boundary_map'
+        # boundary map
+        b_map = DocumentDataAccessServiceV2.create_document(
+            approved_user_operator.user_id, data_url_to_file(MOCK_DATA_URL), 'boundary_map', operation.id
         )
-        pfd = DocumentDataAccessService.create_document(
-            approved_user_operator.user_id, data_url_to_file(MOCK_DATA_URL), 'process_flow_diagram'
+        # process flow diagram
+        DocumentDataAccessServiceV2.create_document(
+            approved_user_operator.user_id, data_url_to_file(MOCK_DATA_URL), 'process_flow_diagram', operation.id
         )
-        operation.documents.set([b_map, pfd])
 
         assert Document.objects.count() == 2
         assert operation.documents.count() == 2
